requirements.txt
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.8
numpy>=1.23.5
opencv-python>=4.6.0
Pillow>=9.3.0
torch>=1.13.0
torchvision>=0.14.0
tqdm>=4.64.1
realesrgan>=0.3.0
runpod>=1.1.0
requests==2.31.0 